backend/requirements.txt
fastapi==0.114.1
uvicorn[standard]==0.30.6
openai>=1.40.0
python-dotenv>=1.0.1

